Rename types.ts to schema.ts
