+++ src/lib/database/index.ts
@@ -1,4 +1,4 @@
 export * from "./sample.ts"
+export * from "./schema.ts"
 export * from "./singleton.ts"
 export * from "./store.ts"
-export * from "./types.ts"
